Refactoring
- Change offersdata reducer name
