+++ src/redux/offers-data/offers-data.js
