+++ src/redux/offers-data/offers-data.test.js
@@ -1,6 +1,6 @@
 import MockAdapter from "axios-mock-adapter";
 import {createAPI} from "../../api/api.js";
-import {reducer, ActionType, ActionCreator, Operation} from "./offersData.js";
+import {reducer, ActionType, ActionCreator, Operation} from "./offers-data.js";
 import {nearOffers} from "../../mocks/near-offers";
 
 const cityOffers = [
