+++ src/redux/offers-data/selectors.js
