+++ src/redux/reducers.js
@@ -1,5 +1,5 @@
 import {combineReducers} from "redux";
-import {reducer as offersData} from "./offersData/offersData.js";
+import {reducer as offersData} from "./offers-data/offers-data.js";
 import {reducer as page} from "./page/page.js";
 import {reducer as user} from "./user/user.js";
 import NameSpace from "./name-space.js";
